move local/domain to private classes in tv
